phase control to the top
